+++ agentloop-20260520/alibabacloud_agentloop20260520/models/__init__.py
@@ -15,8 +15,9 @@ from ._index_key import IndexKey
 from ._model_parameters import ModelParameters
 from ._prompt_template_item import PromptTemplateItem
 from ._run_strategies import RunStrategies
-from ._add_mem_0memories_request import AddMem0MemoriesRequest
-from ._add_mem_0memories_response import AddMem0MemoriesResponse
+from ._add_dataset_data_request import AddDatasetDataRequest
+from ._add_dataset_data_response_body import AddDatasetDataResponseBody
+from ._add_dataset_data_response import AddDatasetDataResponse
 from ._create_agent_space_request import CreateAgentSpaceRequest
 from ._create_agent_space_response_body import CreateAgentSpaceResponseBody
 from ._create_agent_space_response import CreateAgentSpaceResponse
@@ -41,11 +42,6 @@ from ._delete_context_store_apikey_response import DeleteContextStoreAPIKeyRespo
 from ._delete_dataset_request import DeleteDatasetRequest
 from ._delete_dataset_response_body import DeleteDatasetResponseBody
 from ._delete_dataset_response import DeleteDatasetResponse
-from ._delete_mem_0memories_request import DeleteMem0MemoriesRequest
-from ._delete_mem_0memories_shrink_request import DeleteMem0MemoriesShrinkRequest
-from ._delete_mem_0memories_response import DeleteMem0MemoriesResponse
-from ._delete_mem_0memory_request import DeleteMem0MemoryRequest
-from ._delete_mem_0memory_response import DeleteMem0MemoryResponse
 from ._delete_pipeline_request import DeletePipelineRequest
 from ._delete_pipeline_response_body import DeletePipelineResponseBody
 from ._delete_pipeline_response import DeletePipelineResponse
@@ -67,10 +63,6 @@ from ._get_context_store_apikey_response import GetContextStoreAPIKeyResponse
 from ._get_dataset_request import GetDatasetRequest
 from ._get_dataset_response_body import GetDatasetResponseBody
 from ._get_dataset_response import GetDatasetResponse
-from ._get_mem_0memories_request import GetMem0MemoriesRequest
-from ._get_mem_0memories_response import GetMem0MemoriesResponse
-from ._get_mem_0memory_request import GetMem0MemoryRequest
-from ._get_mem_0memory_response import GetMem0MemoryResponse
 from ._get_pipeline_request import GetPipelineRequest
 from ._get_pipeline_response_body import GetPipelineResponseBody
 from ._get_pipeline_response import GetPipelineResponse
@@ -92,8 +84,6 @@ from ._list_pipelines_response import ListPipelinesResponse
 from ._search_context_request import SearchContextRequest
 from ._search_context_response_body import SearchContextResponseBody
 from ._search_context_response import SearchContextResponse
-from ._search_mem_0memories_request import SearchMem0MemoriesRequest
-from ._search_mem_0memories_response import SearchMem0MemoriesResponse
 from ._update_agent_space_request import UpdateAgentSpaceRequest
 from ._update_agent_space_response_body import UpdateAgentSpaceResponseBody
 from ._update_agent_space_response import UpdateAgentSpaceResponse
@@ -103,13 +93,9 @@ from ._update_context_store_response import UpdateContextStoreResponse
 from ._update_dataset_request import UpdateDatasetRequest
 from ._update_dataset_response_body import UpdateDatasetResponseBody
 from ._update_dataset_response import UpdateDatasetResponse
-from ._update_mem_0memory_request import UpdateMem0MemoryRequest
-from ._update_mem_0memory_response import UpdateMem0MemoryResponse
 from ._update_pipeline_request import UpdatePipelineRequest
 from ._update_pipeline_response_body import UpdatePipelineResponseBody
 from ._update_pipeline_response import UpdatePipelineResponse
-from ._validate_mem_0apikey_request import ValidateMem0APIKeyRequest
-from ._validate_mem_0apikey_response import ValidateMem0APIKeyResponse
 from ._create_context_store_request import CreateContextStoreRequestConfigSource
 from ._create_context_store_request import CreateContextStoreRequestConfig
 from ._describe_regions_response_body import DescribeRegionsResponseBodyRegions
@@ -158,8 +144,9 @@ __all__ = [
     ModelParameters,
     PromptTemplateItem,
     RunStrategies,
-    AddMem0MemoriesRequest,
-    AddMem0MemoriesResponse,
+    AddDatasetDataRequest,
+    AddDatasetDataResponseBody,
+    AddDatasetDataResponse,
     CreateAgentSpaceRequest,
     CreateAgentSpaceResponseBody,
     CreateAgentSpaceResponse,
@@ -184,11 +171,6 @@ __all__ = [
     DeleteDatasetRequest,
     DeleteDatasetResponseBody,
     DeleteDatasetResponse,
-    DeleteMem0MemoriesRequest,
-    DeleteMem0MemoriesShrinkRequest,
-    DeleteMem0MemoriesResponse,
-    DeleteMem0MemoryRequest,
-    DeleteMem0MemoryResponse,
     DeletePipelineRequest,
     DeletePipelineResponseBody,
     DeletePipelineResponse,
@@ -210,10 +192,6 @@ __all__ = [
     GetDatasetRequest,
     GetDatasetResponseBody,
     GetDatasetResponse,
-    GetMem0MemoriesRequest,
-    GetMem0MemoriesResponse,
-    GetMem0MemoryRequest,
-    GetMem0MemoryResponse,
     GetPipelineRequest,
     GetPipelineResponseBody,
     GetPipelineResponse,
@@ -235,8 +213,6 @@ __all__ = [
     SearchContextRequest,
     SearchContextResponseBody,
     SearchContextResponse,
-    SearchMem0MemoriesRequest,
-    SearchMem0MemoriesResponse,
     UpdateAgentSpaceRequest,
     UpdateAgentSpaceResponseBody,
     UpdateAgentSpaceResponse,
@@ -246,13 +222,9 @@ __all__ = [
     UpdateDatasetRequest,
     UpdateDatasetResponseBody,
     UpdateDatasetResponse,
-    UpdateMem0MemoryRequest,
-    UpdateMem0MemoryResponse,
     UpdatePipelineRequest,
     UpdatePipelineResponseBody,
     UpdatePipelineResponse,
-    ValidateMem0APIKeyRequest,
-    ValidateMem0APIKeyResponse,
     CreateContextStoreRequestConfigSource,
     CreateContextStoreRequestConfig,
     DescribeRegionsResponseBodyRegions,